LPS-71243 move ModularTrash* from service.wrapper to service
